Moved instruction panel
