+++ src/index.js
@@ -4,90 +4,13 @@ import { datGUI } from './dat';
 import { initVisibilityDesider } from './visibiltyDesider';
 import { loadUniverseAt, unLoadUniverseAt, updateloadedParts, initMaping } from './objectManager'
 import { initController, updateController } from './controller'
+import instructionPanel  from './instructionPanel'
 import setLocationHash from 'set-location-hash';
 
 
 
 
 var camera, scene, renderer, controls;
-
-
-var blocker = document.getElementById('blocker');
-var instructions = document.getElementById('instructions');
-
-//(function () {
-    // http://www.html5rocks.com/en/tutorials/pointerlock/intro/
-
-    var havePointerLock = 'pointerLockElement' in document || 'mozPointerLockElement' in document || 'webkitPointerLockElement' in document;
-
-    if (havePointerLock) {
-
-        var element = document.body;
-
-        var pointerlockchange = function (event) {
-
-            if (document.pointerLockElement === element || document.mozPointerLockElement === element || document.webkitPointerLockElement === element) {
-
-
-                controls.enabled = true;
-
-                blocker.style.display = 'none';
-
-            } else {
-
-                controls.enabled = false;
-
-                blocker.style.display = 'block';
-
-                instructions.style.display = '';
-
-            }
-
-        };
-
-
-
-        var pointerlockerror = function (event) {
-
-            instructions.style.display = '';
-
-        };
-
-        // Hook pointer lock state change events
-        document.addEventListener('pointerlockchange', pointerlockchange, false);
-        document.addEventListener('mozpointerlockchange', pointerlockchange, false);
-        document.addEventListener('webkitpointerlockchange', pointerlockchange, false);
-
-        document.addEventListener('pointerlockerror', pointerlockerror, false);
-        document.addEventListener('mozpointerlockerror', pointerlockerror, false);
-        document.addEventListener('webkitpointerlockerror', pointerlockerror, false);
-
-
-        function firsttlock() {
-            element.requestPointerLock();
-            element.removeEventListener('click', firsttlock);
-        }
-
-        element.addEventListener('click', firsttlock);
-
-
-        instructions.addEventListener('click', function (event) {
-
-            instructions.style.display = 'none';
-
-            // Ask the browser to lock the pointer
-            element.requestPointerLock = element.requestPointerLock || element.mozRequestPointerLock || element.webkitRequestPointerLock;
-            element.requestPointerLock();
-
-        }, false);
-
-    } else {
-
-        instructions.innerHTML = 'Your browser doesn\'t seem to support Pointer Lock API';
-
-    }
-
-//})();
 
 
 
@@ -200,11 +123,10 @@ function init(position) {
     // scene.add(light);
     controls = initController(camera)
     scene.add(controls.getObject());
-
-
     camera.position.y = 100;
     controls.getObject().translateX(position.x);
     controls.getObject().translateZ(position.z);
+    instructionPanel.init(controls);
     // let lastCameraRotation =localStorage.getItem("lastCameraRotation");
     // if (lastCameraRotation) {
     //     let obj = JSON.parse(lastCameraRotation);
