+++ src/instructionPanel.js
@@ -1,0 +1,90 @@
+
+
+
+let panel = {
+
+init : function (controls) {
+    
+var blocker = document.getElementById('blocker');
+var instructions = document.getElementById('instructions');
+
+
+        // http://www.html5rocks.com/en/tutorials/pointerlock/intro/
+
+    var havePointerLock = 'pointerLockElement' in document || 'mozPointerLockElement' in document || 'webkitPointerLockElement' in document;
+
+    if (havePointerLock) {
+
+        var element = document.body;
+
+        var pointerlockchange = function (event) {
+
+            if (document.pointerLockElement === element || document.mozPointerLockElement === element || document.webkitPointerLockElement === element) {
+
+
+                controls.enabled = true;
+
+                blocker.style.display = 'none';
+
+            } else {
+
+                controls.enabled = false;
+
+                blocker.style.display = 'block';
+
+                instructions.style.display = '';
+
+            }
+
+        };
+
+
+
+        var pointerlockerror = function (event) {
+
+            instructions.style.display = '';
+
+        };
+
+        // Hook pointer lock state change events
+        document.addEventListener('pointerlockchange', pointerlockchange, false);
+        document.addEventListener('mozpointerlockchange', pointerlockchange, false);
+        document.addEventListener('webkitpointerlockchange', pointerlockchange, false);
+
+        document.addEventListener('pointerlockerror', pointerlockerror, false);
+        document.addEventListener('mozpointerlockerror', pointerlockerror, false);
+        document.addEventListener('webkitpointerlockerror', pointerlockerror, false);
+
+
+        function firsttlock() {
+            element.requestPointerLock();
+            element.removeEventListener('click', firsttlock);
+        }
+
+        element.addEventListener('click', firsttlock);
+
+
+        instructions.addEventListener('click', function (event) {
+
+            instructions.style.display = 'none';
+
+            // Ask the browser to lock the pointer
+            element.requestPointerLock = element.requestPointerLock || element.mozRequestPointerLock || element.webkitRequestPointerLock;
+            element.requestPointerLock();
+
+        }, false);
+
+    } else {
+
+        instructions.innerHTML = 'Your browser doesn\'t seem to support Pointer Lock API';
+
+    }
+}
+
+
+
+
+
+}
+
+export default panel;
